Organize runtime internals
